ultimate-pro-quality/requirements.txt
python-vlc==3.0.18122
screeninfo~=0.8.1
